+++ src/SMAPI.Toolkit/Framework/Clients/ApiCacheHeaders.cs
@@ -13,7 +13,7 @@ namespace StardewModdingAPI.Toolkit.Framework.Clients
         public readonly DateTimeOffset LastModified;
 
         /// <summary>The entity tag which represents the current version of the server's data.</summary>
-        public readonly string EntityTag;
+        public readonly string? EntityTag;
 
 
         /*********
@@ -22,7 +22,7 @@ namespace StardewModdingAPI.Toolkit.Framework.Clients
         /// <summary>Construct an instance.</summary>
         /// <param name="lastModified">When the server's data was last updated.</param>
         /// <param name="entityTag">The entity tag which represents the current version of the server's data.</param>
-        public ApiCacheHeaders(DateTimeOffset lastModified, string entityTag)
+        public ApiCacheHeaders(DateTimeOffset lastModified, string? entityTag)
         {
             this.LastModified = lastModified;
             this.EntityTag = entityTag;
@@ -30,12 +30,18 @@ namespace StardewModdingAPI.Toolkit.Framework.Clients
 
         /// <summary>Read the required cache headers from an API response.</summary>
         /// <param name="response">The API response whose headers to read.</param>
+        /// <param name="allowNullEntityTag">Whether to ignore a missing <c>ETag</c> header instead of throwing an exception.</param>
         /// <exception cref="InvalidOperationException">The response is missing one or more of the required HTTP headers (<c>ETag</c> and <c>Last-Modified</c>).</exception>
-        public static ApiCacheHeaders FromResponse(IResponse response)
+        public static ApiCacheHeaders FromResponse(IResponse response, bool allowNullEntityTag = false)
         {
             return new ApiCacheHeaders(
                 lastModified: response.Message.Content.Headers.LastModified ?? throw new InvalidOperationException("The API response doesn't include the required Last-Modified header."),
-                entityTag: response.Message.Headers.ETag?.Tag ?? throw new InvalidOperationException("The API response doesn't include the required ETag header.")
+                entityTag:
+                    response.Message.Headers.ETag?.Tag
+                    ?? (allowNullEntityTag
+                        ? null
+                        : throw new InvalidOperationException("The API response doesn't include the required ETag header.")
+                    )
             );
         }
     }