+++ docs/release-notes.md
@@ -19,8 +19,8 @@
   * Fixed player sprites and building paint masks not always propagated on change.
 
 * For the update check server:
-  * Rewrote update checks for mods on CurseForge to use a new CurseForge API endpoint.  
-    _This should result in much faster update checks for CurseForge, and less chance of update-check errors when the CurseForge servers are under heavy load._
+  * Rewrote update checks for mods on CurseForge and ModDrop to use new export API endpoints.  
+    _This should result in much faster update checks for those sites, and less chance of update-check errors when their servers are under heavy load._
 
 * For the log parser:
   * Clicking a checkbox in the mod list now always only changes that checkbox, to allow hiding a single mod.